P7.1 — Five nav sections, and every old URL still resolves (#328)
The nav carried sixteen top-level items. Shopify's sidebar collapses past seven,
so a third of the app sat behind "View more" on a normal screen -- built, paid
for, and undiscoverable. It is five now: Home, Campaigns, Prices, Imports,
Settings.

Nothing is deleted and nothing is redesigned. Every route file moves unchanged;
the section layouts add a plain sub-navigation and an Outlet. Deliberately plain
links rather than tabs -- the tabbed treatment belongs with each section's
redesign, and shipping half of it here would leave two navigation idioms in the
app at once, which is the problem this epic exists to fix, in miniature.

Prices gathers catalogue, baselines, costs, what-is-live and drift: five tables
over the same variant rows, 997 lines to show one thing with different columns.
Imports gathers the three CSV routes plus recapture -- one verb that had three
destinations. Settings gathers segments, plan, feedback and diagnostics.

The awkward one is `/app/prices/import`, which now sits under the prices section
it used to sound like it belonged to, so it redirects out of its own subtree
rather than being shadowed by it.

`/app/activity` keeps its URL and leaves the nav; folding it into Home needs the
tabs from P7.10. `/app/calendar` lands on `/app/campaigns/calendar` rather than
`?view=calendar`, because the query-param toggle means merging two route modules'
loaders, which is P7.5's job. Both tickets updated.

**The old URLs are not ours to break.** They are linked from operator alerts,
from runbooks somebody opens while a run is misbehaving, and from whatever a
merchant bookmarked -- `/app/debug` 404ing at the moment it is most needed is the
failure worth guarding. `legacy-routes.ts` records the promise and its test holds
it: every old path has a redirect, every destination has a route file to land on,
no redirect points at another redirect, and nothing inside the app links through
a redirect rather than at the real page. All of it reads the routes directory
rather than a second hand-kept list, because two hand-kept lists agreeing with
each other is the contract-drift shape this codebase keeps paying for.

The redirect stubs authenticate before redirecting. The compliance suite caught
that they did not, which was right: an embedded route reached without a session
has to go through OAuth, and doing it here means the merchant lands on the page
they asked for instead of bouncing off a second redirect.

Two existing tests also caught the moves -- `display.test.ts` named two routes by
path. Updated rather than loosened.
